requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
croniter==2.0.2
frozenlist==1.4.1
hiredis==2.3.2
idna==3.6
multidict==6.0.5
pyTelegramBotAPI==4.16.1
python-cron==1.0.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.2
requests==2.31.0
six==1.16.0
urllib3==2.2.1
yarl==1.9.4
